repository updates

+++ src/main/kotlin/com/miguel/game/bank/Currency.kt
@@ -4,12 +4,12 @@ import org.bukkit.Material
 
 interface ICurrency {
     val material: Material
-    val value: Float
+    val value: Double
 }
 
 data class Currency(
     override val material: Material,
-    override val value: Float
+    override val value: Double
 ) : ICurrency
 
 data class Amount(
